dom/media/test/test_cloneElementVisually_ended_video.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test cloneElementVisually</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="https://example.com:443/tests/dom/media/test/cloneElementVisually_helpers.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
</head>
<body>
<div id="content">
  <h1>Original</h1>
  <video id="original"></video>
  <h1>Clone</h1>
</div>
<div id="results">
  <h1>Results</h1>
  <canvas id="left"></canvas>
  <canvas id="right"></canvas>
</div>

<script type="application/javascript">

/**
 * Test that when we start cloning a video that has already ended, the
 * clone displays the last frame from the video.
 */
add_task(async () => {
  await setup();

  let originalVideo = document.getElementById("original");
  let ended = waitForEventOnce(originalVideo, "ended");
  await originalVideo.play();
  await ended;

  await withNewClone(originalVideo, async clone => {
    SpecialPowers.wrap(originalVideo).cloneElementVisually(clone);

    ok(await assertVideosMatch(originalVideo, clone),
       "Visual clone should display final frame.");
  });
});

</script>

</body>
</html>
